+++ planner/cardinality/row_count_index.go
@@ -98,7 +98,7 @@ func GetIndexRowCountForStatsV1(sctx sessionctx.Context, coll *statistics.HistCo
 	totalCount := float64(0)
 	for _, ran := range indexRanges {
 		if debugTrace {
-			statistics.DebugTraceStartEstimateRange(sctx, ran, nil, nil, totalCount)
+			DebugTraceStartEstimateRange(sctx, ran, nil, nil, totalCount)
 		}
 		rangePosition := statistics.GetOrdinalOfRangeCond(sc, ran)
 		var rangeVals []types.Datum
@@ -118,7 +118,7 @@ func GetIndexRowCountForStatsV1(sctx sessionctx.Context, coll *statistics.HistCo
 				return 0, errors.Trace(err)
 			}
 			if debugTrace {
-				statistics.DebugTraceEndEstimateRange(sctx, count, statistics.DebugTraceRange)
+				DebugTraceEndEstimateRange(sctx, count, DebugTraceRange)
 			}
 			totalCount += count
 			continue
@@ -185,7 +185,7 @@ func GetIndexRowCountForStatsV1(sctx sessionctx.Context, coll *statistics.HistCo
 		}
 		count := selectivity * idx.TotalRowCount()
 		if debugTrace {
-			statistics.DebugTraceEndEstimateRange(sctx, count, statistics.DebugTraceRange)
+			DebugTraceEndEstimateRange(sctx, count, DebugTraceRange)
 		}
 		totalCount += count
 	}
@@ -230,14 +230,14 @@ func GetIndexRowCountForStatsV2(sctx sessionctx.Context, idx *statistics.Index, 
 			return 0, err
 		}
 		if debugTrace {
-			statistics.DebugTraceStartEstimateRange(sctx, indexRange, lb, rb, totalCount)
+			DebugTraceStartEstimateRange(sctx, indexRange, lb, rb, totalCount)
 		}
 		fullLen := len(indexRange.LowVal) == len(indexRange.HighVal) && len(indexRange.LowVal) == len(idx.Info.Columns)
 		if bytes.Equal(lb, rb) {
 			// case 1: it's a point
 			if indexRange.LowExclude || indexRange.HighExclude {
 				if debugTrace {
-					statistics.DebugTraceEndEstimateRange(sctx, 0, statistics.DebugTraceImpossible)
+					DebugTraceEndEstimateRange(sctx, 0, DebugTraceImpossible)
 				}
 				continue
 			}
@@ -246,7 +246,7 @@ func GetIndexRowCountForStatsV2(sctx sessionctx.Context, idx *statistics.Index, 
 				if idx.Info.Unique {
 					totalCount++
 					if debugTrace {
-						statistics.DebugTraceEndEstimateRange(sctx, 1, statistics.DebugTraceUniquePoint)
+						DebugTraceEndEstimateRange(sctx, 1, DebugTraceUniquePoint)
 					}
 					continue
 				}
@@ -254,7 +254,7 @@ func GetIndexRowCountForStatsV2(sctx sessionctx.Context, idx *statistics.Index, 
 				// If the current table row count has changed, we should scale the row count accordingly.
 				count *= idx.GetIncreaseFactor(realtimeRowCount)
 				if debugTrace {
-					statistics.DebugTraceEndEstimateRange(sctx, count, statistics.DebugTracePoint)
+					DebugTraceEndEstimateRange(sctx, count, DebugTracePoint)
 				}
 				totalCount += count
 				continue
@@ -327,7 +327,7 @@ func GetIndexRowCountForStatsV2(sctx sessionctx.Context, idx *statistics.Index, 
 		}
 
 		if debugTrace {
-			statistics.DebugTraceEndEstimateRange(sctx, count, statistics.DebugTraceRange)
+			DebugTraceEndEstimateRange(sctx, count, DebugTraceRange)
 		}
 		totalCount += count
 	}
